add lib colisoes
